# faster internal group_keys()