--- a/R/group_keys.R
+++ b/R/group_keys.R
@@ -1,10 +1,3 @@
-group_keys_impl <- function(.data) {
-  structure(
-    select(group_data(.data), -last_col()),
-    .drop = NULL
-  )
-}
-
 #' @rdname group_split
 #' @export
 group_keys <- function(.tbl, ...) {
@@ -13,7 +6,7 @@ group_keys <- function(.tbl, ...) {
 
 #' @export
 group_keys.data.frame <- function(.tbl, ...){
-  group_keys_impl(group_by(.tbl, ...))
+  .Call(`dplyr_group_keys_impl`, group_by(.tbl, ...))
 }
 
 #' @export
@@ -21,7 +14,7 @@ group_keys.grouped_df <- function(.tbl, ...) {
   if (dots_n(...)) {
     warn("... is ignored in group_keys(<grouped_df>), please use group_by(..., add = TRUE) %>% group_keys()")
   }
-  group_keys_impl(.tbl)
+  .Call(`dplyr_group_keys_impl`, .tbl)
 }
 
 #' @export