--- a/src/dplyr/symbols.h
+++ b/src/dplyr/symbols.h
@@ -12,6 +12,7 @@ struct symbols {
 
 struct vectors {
   static SEXP classes_vctrs_list_of;
+  static SEXP classes_tbl_df;
   static SEXP empty_int_vector;
 };