``` r
library(bench)
library(dplyr)

callr::r(function(){
  library(dplyr)
  df <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100))) %>% group_by(g)
  bench::mark(a = group_keys(df))
}, lib = "/Users/romainfrancois/git/tidyverse/dplyr/bench-libs/0.8.3")
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 a             721µs    774µs     1217.     515KB     51.2

callr::r(function(){
  library(dplyr)
  df <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100))) %>% group_by(g)
  bench::mark(a = group_keys(df))
}, lib = "/Users/romainfrancois/git/tidyverse/dplyr/bench-libs/master")
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 a             1.3ms   1.44ms      663.    70.8KB     50.7

callr::r(function(){
  library(dplyr)
  df <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100))) %>% group_by(g)
  bench::mark(group_keys(df))
})
#> # A tibble: 1 x 6
#>   expression          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 group_keys(df)   2.18µs   2.55µs   375549.    7.22KB     75.1
```

<sup>Created on 2019-12-04 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>